Require blank after namelist group name name & update tests
Per F2008 section 10.11.3.2: "A namelist comment is also permitted to start in
the first NONBLANK position of an input record..."; and section 10.11.3.1, part
(3), "ONE OR MORE blanks" (emphasis added). That is, the namelist group name
must be followed by one or more blanks, meaning that the position immeidately
following the namelist group name is NOT nonblank, and therefore may not contain
a namelist comment.